+++ lib/titan.rb
@@ -2,5 +2,6 @@ require "titan/version"
 
 module Titan
   autoload :CLI,      "titan/cli"
+  autoload :System,   "titan/system"
   autoload :Thread,   "titan/thread"
 end